Port to ABI3

+++ csrc/fused/fused.h
@@ -1,15 +1,14 @@
-
-#include <torch/extension.h>
+#include <torch/all.h>
 
 void transpose_pad_permute_cuda(
                 torch::Tensor input,
                 torch::Tensor output,
-                int tensor_layout);
+                int64_t tensor_layout);
 
 void scale_fuse_quant_cuda(
                 torch::Tensor input,
                 torch::Tensor output,
                 torch::Tensor scale,
-                int num_tokens,
-                float scale_max,
-                int tensor_layout);
+                int64_t num_tokens,
+                double scale_max,
+                int64_t tensor_layout);
